provide the meta method to the DraftStorage
