+++ libs/newsletters-data-client/src/lib/draft-storage/DraftStorage.ts
@@ -1,9 +1,16 @@
-import type { DraftNewsletterData } from '../newsletter-data-type';
+import type { MetaData } from '../meta-data-type';
+import { createNewMeta, stripMeta, updateMeta } from '../meta-data-type';
+import type {
+	DraftNewsletterData,
+	DraftNewsletterDataWithMeta,
+	DraftNewsletterDataWithoutMeta,
+} from '../newsletter-data-type';
 import { isDraftNewsletterData } from '../newsletter-data-type';
 import type {
 	SuccessfulStorageResponse,
 	UnsuccessfulStorageResponse,
 } from '../storage-response-types';
+import type { UserProfile } from '../user-profile';
 
 export type DraftWithoutId = DraftNewsletterData & { listId: undefined };
 export type DraftWithId = DraftNewsletterData & { listId: number };
@@ -38,4 +45,18 @@ export abstract class DraftStorage {
 	abstract readAll(): Promise<
 		SuccessfulStorageResponse<DraftWithId[]> | UnsuccessfulStorageResponse
 	>;
+
+	stripMeta(
+		data: DraftNewsletterDataWithMeta | DraftNewsletterData,
+	): DraftNewsletterDataWithoutMeta {
+		return stripMeta(data);
+	}
+
+	createNewMeta(user: UserProfile): MetaData {
+		return createNewMeta(user);
+	}
+
+	updateMeta(meta: MetaData, user: UserProfile): MetaData {
+		return updateMeta(meta, user);
+	}
 }
