+++ libs/newsletters-data-client/src/lib/meta-data-type.ts
@@ -1,4 +1,5 @@
-import { z } from 'zod';
+import { unknown, z } from 'zod';
+import type { UserProfile } from './user-profile';
 
 export const metaDataSchema = z.object({
 	createdTimestamp: z.number(),
@@ -8,3 +9,31 @@ export const metaDataSchema = z.object({
 });
 
 export type MetaData = z.infer<typeof metaDataSchema>;
+
+export const createNewMeta = (user: UserProfile): MetaData => {
+	const now = Date.now();
+	return {
+		createdTimestamp: now,
+		createdBy: user.email ?? '[unknown]',
+		updatedTimestamp: now,
+		updatedBy: user.email ?? '[unknown]',
+	};
+};
+
+export const updateMeta = (meta: MetaData, user: UserProfile): MetaData => {
+	const now = Date.now();
+	return {
+		...meta,
+		updatedTimestamp: now,
+		updatedBy: user.email ?? '[unknown]',
+	};
+};
+
+export const stripMeta = <T extends Partial<Record<string, unknown>>>(
+	data: T,
+) => {
+	return {
+		...data,
+		meta: undefined,
+	};
+};
